requirements.txt
# 🟢 Django Framework + Extensions
Django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0

# 🟢 Environment Variables
python-dotenv==1.1.1

# 🟢 WSGI Server (nur für Produktion nötig)
gunicorn==23.0.0

# 🟢 JWT Token Handling
PyJWT==2.9.0

# 🟢 Sicherheits-Utilities
itsdangerous==2.2.0

# 🟢 Django interne Abhängigkeiten
sqlparse==0.5.3
asgiref==3.8.1
tzdata==2025.2
packaging==25.0

# 🟢 PostgreSQL Treiber (für AWS Produktion)
psycopg2-binary==2.9.10
